Problem
Trying to add more mob texture variants for more mobs: currently the iron golem and the slime.

I've been successful but I have a problem: A "missing texture" variant is also being added to the texture variant pool. if the iron golem count is set to 0, then only the missing texture variant will show up on iron golems, if set to 1, then one of the textures i've added OR the missing texture variant will show up on iron golems. The same happens for Slimes.